Move charset tests to new file, add new ch function for get the text height also
